requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
by==0.0.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
et-xmlfile==1.1.0
Faker==8.12.1
idna==2.10
iniconfig==1.1.1
openpyxl==3.0.7
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
requests==2.25.1
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
Appium-Python-Client~=1.2.0